# gi_service/contract/rest/api_contract.yaml
openapi: 3.1.0
info:
  title: OpenGIN Service API
  description: This document gives the structure of APIs for the OpenGIN Services.
  version: 1.0.0
servers:
- url: http://localhost:8081
...
paths:
...
  /departments-by-portfolio/{portfolioId}:
    post:
      tags:
      - Organization
      summary: Get active departments for a portfolio.
      description: Returns a list of departments under a given portfolio and a given date.
      operationId: getDepartmentsByPortfolio
      parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrive the list of active departments.
        content:
          application/json:
            schema:
              type: object
              description: Required date.
              properties:
                activeDate:
                  type: object
                  required:
                  - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    properties:
                      totalDepartments:
                        type: integer
                      newDepartments:
                        type: integer
                      departmentList:
                        type: array
                        items:
                          type: object
                          required:
                          - name
                          - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            isNew:
                              type: boolean
                            hasData:
                              type: boolean
              
              example:
                body:
                - totalDepartments: 2
                  newDepartments: 2
                  departmentList:
                  - name: National Police Academy
                    id: 2187-27_dep_4
                    isNew: false
                    hasData: true
                  - name: Assets vested to the Secretary to the Treasury, under the Act, No. 43 of 2011
                    id: 2153-12_dep_437
                    isNew: false
                    hasData: false
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error
  /persons-by-portfolio/{portfolioId}:
    post:
      tags:
      - Organization
      summary: Get active persons for a portfolio.
      description: Returns a list of persons under a given portfolio and a given date.
      operationId: getPersonsByMinistry
      parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrieve the list of active persons.
        content:
          application/json:
            schema:
              type: object
              description: required date
              properties:
                presidentId:
                  type: string
                activeDate:
                  type: object
                  required:
                  - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    properties:
                      totalPeople:
                        type: integer
                      peopleList:
                        type: array
                        items:
                          type: object
                          required:
                          - name
                          - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            isPresident:
                              type: boolean
                            isNew:
                              type: boolean

              example:
                body:
                - totalPeople:
                  peopleList:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03_cit_1
                    isPresident: true
                    isNew: false
                  - name: Gotabaya Rajapaksa
                    id: 2149-34_cit_1
                    isPresident: false
                    isNew: false
                  
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error
...